pyproject.toml
[project]
name = "exdb-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "John Faucett", email = "jsfaucett87@icloud.com" }
]
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.2",
    "owlready2>=0.47",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
exdb-data = "exdb_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
